Move sample to a different package
