+++ ComposeTesting/sample/src/androidTest/java/com/microsoft/device/dualscreen/testing/sample/FoldingFeatureHelperTest.kt
@@ -5,12 +5,17 @@
  *
  */
 
-package com.microsoft.device.dualscreen.testing
+package com.microsoft.device.dualscreen.testing.sample
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import com.microsoft.device.dualscreen.testing.ui.theme.ComposeTestingTheme
+import com.microsoft.device.dualscreen.testing.createWindowLayoutInfoPublisherRule
+import com.microsoft.device.dualscreen.testing.sample.R
+import com.microsoft.device.dualscreen.testing.getString
+import com.microsoft.device.dualscreen.testing.simulateHorizontalFoldingFeature
+import com.microsoft.device.dualscreen.testing.simulateVerticalFoldingFeature
+import com.microsoft.device.dualscreen.testing.sample.ui.theme.ComposeTestingTheme
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
