+++ ComposeTesting/sample/src/main/java/com/microsoft/device/dualscreen/testing/sample/MainActivity.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.microsoft.device.dualscreen.testing
+package com.microsoft.device.dualscreen.testing.sample
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -15,7 +15,7 @@ import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
-import com.microsoft.device.dualscreen.testing.ui.theme.ComposeTestingTheme
+import com.microsoft.device.dualscreen.testing.sample.ui.theme.ComposeTestingTheme
 import com.microsoft.device.dualscreen.twopanelayout.TwoPaneLayout
 
 class MainActivity : ComponentActivity() {
