+++ ComposeTesting/sample/src/main/java/com/microsoft/device/dualscreen/testing/sample/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.dualscreen.testing.ui.theme
+package com.microsoft.device.dualscreen.testing.sample.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
