+++ ComposeTesting/sample/src/main/java/com/microsoft/device/dualscreen/testing/sample/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.dualscreen.testing.ui.theme
+package com.microsoft.device.dualscreen.testing.sample.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
