+++ ComposeTesting/sample/src/main/java/com/microsoft/device/dualscreen/testing/sample/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.dualscreen.testing.ui.theme
+package com.microsoft.device.dualscreen.testing.sample.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
